gateway/requirements.txt
# SentinelLM Gateway — Python Dependencies

# --- Web Framework ---
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# --- HTTP Client ---
httpx>=0.27.0

# --- Database ---
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# --- Configuration ---
pydantic-settings>=2.0.0
pyyaml>=6.0

# --- PII Detection (Presidio) ---
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# --- Semantic Model (DistilBERT) ---
transformers>=4.36.0
torch>=2.1.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx  # for FastAPI TestClient

# --- Linting / Type Checking ---
ruff>=0.1.0
mypy>=1.7.0

# --- ML Evaluation ---
seqeval>=1.2.0
